refactor(griditem): handle pr feedback moving things around
