+++ package.json
@@ -32,7 +32,7 @@
     "clean": "rimraf './dist'",
     "clean:stories": "rimraf './storybook-static'",
     "format": "prettier . --write",
-    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0 && npm run lint:styles",
+    "lint": "tsc && eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0 && npm run lint:styles",
     "lint:styles": "stylelint \"src/**/*.scss\" --fix",
     "prepare": "husky install"
   },
