+++ src/components/GridItem/types.ts
@@ -1,0 +1,5 @@
+export enum GridItemAlign {
+  center = 'center',
+  left = 'left',
+  right = 'right',
+}
